wip: simplify server translations for original text
We can inline the original text at least in dev mode. If can skip metadata being used it's good.
If we want to keep the bundle as small as possible, then we can probably ask fro english translations as well, which should be as the separate file and through a common interface via language provider/fetch